fix(entities): change all Optional[int] to Optional[int64] across hand-written entities
Picklist and integer fields in Ticket, Company, Project, Task, Contract,
ConfigurationItem, and TicketNote used Optional[int] but the Autotask API
returns int64 values. This caused JSON decode failures on some platforms
and was inconsistent with generated entities which correctly use int64.

Affected fields: Status, Priority, Source, IssueType, SubIssueType,
TicketType, TicketCategory, CompanyType, Classification, ContractType,
ConfigurationItemType, NoteType, Publish, Type.

+++ crud_entity_test.go
@@ -241,7 +241,7 @@ func TestTicketCRUD(t *testing.T) {
 
 	t.Run("Update", func(t *testing.T) {
 		updated := ticket
-		updated.Priority = autotask.Set(3)
+		updated.Priority = autotask.Set(int64(3))
 		result, err := autotask.Update(t.Context(), client, &updated)
 		if err != nil {
 			t.Fatal(err)
